Lab 09: Added files to work on for subtasks implementation
